New logger: JSON diffs
This is a new logger that outputs the build state in the form of a
series of RFC 6902 JSON patches. The first line it outputs is an empty
object, and then it periodically outputs diffs which transform the
initial object until the build completes.

This is used to efficiently support downstream tools which want to
consume Nix build information and display it in a UI.